% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_content.R
\name{aa_content}
\alias{aa_content}
\title{aa_content}
\usage{
aa_content(peptide)
}
\arguments{
\item{peptide}{string of amino acids}
}
\value{
Bar-plot of amino acid counts in the peptide
}
\description{
aa_content
}
\examples{
aa_content("YGCTRAVSGGTCGCTRACCSAGGTCGGCTRC")
}
