% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{Data in the gene2protein package}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://r4bds.github.io/lab08.html}
}
\usage{
codon_table
}
\description{
The data used in the gene2protein package
}
\keyword{datasets}
