% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons.R
\name{codons}
\alias{codons}
\title{Printing codons for a sequence given}
\usage{
codons(sequence, start = 1)
}
\arguments{
\item{sequence}{any string given will work, but this function is primarily relevant for DNA and protein sequences.}

\item{start}{from which position the function will start returning substrings. Default = 1}
}
\value{
returning codons (substrings of three characters) of a sequence
}
\description{
Printing codons for a sequence given
}
\examples{
codons("ATCGGCTGGCAT")
}
