% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_peptide.R
\name{random_peptide}
\alias{random_peptide}
\title{random_peptide}
\usage{
random_peptide(sequence)
}
\arguments{
\item{sequence}{param_name}
}
\value{
return description of function output
}
\description{
random_peptide
}
\examples{
sequence(1234)
}
