% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_to_aa.R
\name{codon_to_aa}
\alias{codon_to_aa}
\title{codon_to_aa}
\usage{
codon_to_aa(codons)
}
\arguments{
\item{codons}{vector of rna codons}
}
\value{
string of translated amino acid sequence
}
\description{
codon_to_aa
}
\examples{
codon_to_aa("AUG")
}
