% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_T_to_U.R
\name{convert_t_to_u}
\alias{convert_t_to_u}
\title{Substitute thymine with uracil}
\usage{
convert_t_to_u(dna_seq)
}
\arguments{
\item{dna_seq}{text}
}
\value{
rna_seq text
}
\description{
Substitute thymine with uracil
}
\examples{
convert_t_to_u("ATGCTAGCTA")
}
